% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_NNLS.R
\name{deconvolute_NNLS}
\alias{deconvolute_NNLS}
\title{Non-negative Least Squares (NNLS) deconvolution}
\usage{
deconvolute_NNLS(cluster_markers, spots, L1 = 0.01)
}
\arguments{
\item{cluster_markers}{matrix of marker genes vs. clusters, UMI counts}

\item{spots}{matrix of marker genes vs. spots, same dimensions as \code{cluster_markers}}

\item{L1}{L1/LASSO penalty to increase sparsity of the deconvolution}
}
\description{
Use NNLS to deconvolute a matrix of spots consisting of unknown cell types against a reference of known cell type marker gene signatures.
}
\details{
We will need to find an L1 penalty that works best for deconvolution.
}
