% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{find_markers}
\alias{find_markers}
\title{Identify Marker Genes}
\usage{
find_markers(df)
}
\arguments{
\item{df}{dataframe with your counts data}

\item{metadata}{Dataset of metadata}

\item{min_spot_size}{Minimum size of spot desired}

\item{max_spot_size}{Maximum size of spot desired}
}
\description{
identifies marker genes from atlas data
}
