% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spots.R
\name{generate_spots}
\alias{generate_spots}
\title{Generate random spots of given max and min quantity size}
\usage{
generate_spots(df, metadata, min_spot_size = 2, max_spot_size = 8)
}
\arguments{
\item{df}{A dataframe you are generating spot from}

\item{metadata}{Dataset of metadata}

\item{min_spot_size}{Minimum size of spot desired}

\item{max_spot_size}{Maximum size of spot desired}
}
\description{
generate random spots
}
