% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_column_names.R
\name{marker_column_names}
\alias{marker_column_names}
\title{Add column names to Makers Dataset}
\usage{
marker_column_names(metadata, clusters, df)
}
\arguments{
\item{metadata}{original metadata}

\item{clusters}{MUST BE IN "". Name of cluster}

\item{df}{marker dataset}
}
\description{
adding columns
}
