% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_expression.R
\name{subset_expression}
\alias{subset_expression}
\title{Subset the counts data by the Markers Threshold, then ensure expression levels are equal for all columns}
\usage{
subset_expression(markers_threshold, counts, Value)
}
\arguments{
\item{markers_threshold}{Your markers_threshold dataset generated from Identify Marker Thresholding function}

\item{counts}{Your- raw counts matrix}

\item{Value}{Your threshold for how close you want your total number of transcripts to be to the minimum. E.G. all columns within 150% of the minimum, enter 1.5.}
}
\description{
subsetting data
}
