% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding_markers.R
\name{thresholding_markers}
\alias{thresholding_markers}
\title{Marker Thresholding function}
\usage{
thresholding_markers(markers, threshold)
}
\arguments{
\item{markers}{Your Markers dataset generated from Identify Marker Genes function}

\item{theshold}{Your threshold for a marker. Must be a postive integer.}
}
\description{
threshold for acceptable markers
}
